# src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"TripScheduleResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"title": {
						"type": "string"
					},
					"destination": {
						"type": "string"
					},
					"destinationType": {
						"type": "string"
					},
					"startDate": {
						"type": "string"
					},
					"endDate": {
						"type": "string"
					},
					"createdBy": {
						"type": "string"
					},
					"members": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"title",
					"destination",
					"destinationType",
					"startDate",
					"endDate",
					"createdBy",
					"members"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_TripScheduleResponse_": {
				"description": "통일된 API 응답 형식",
				"properties": {
					"isSuccess": {
						"type": "boolean"
					},
					"code": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"allOf": [
							{
								"$ref": "#/components/schemas/TripScheduleResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"isSuccess",
					"code",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateTripRequest": {
				"properties": {
					"title": {
						"type": "string"
					},
					"destination": {
						"type": "string"
					},
					"destinationType": {
						"type": "string"
					},
					"startDate": {
						"type": "string"
					},
					"endDate": {
						"type": "string"
					},
					"members": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"createdBy": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"destination",
					"destinationType",
					"startDate",
					"endDate",
					"members",
					"createdBy"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UpdateTripRequest": {
				"properties": {
					"title": {
						"type": "string"
					},
					"destination": {
						"type": "string"
					},
					"destinationType": {
						"type": "string"
					},
					"startDate": {
						"type": "string"
					},
					"endDate": {
						"type": "string"
					},
					"members": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"createdBy": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"destination",
					"destinationType",
					"startDate",
					"endDate",
					"members",
					"createdBy"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Cost": {
				"properties": {
					"category": {
						"type": "string"
					},
					"value": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"category",
					"value"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TripEventResponse": {
				"properties": {
					"tripId": {
						"type": "number",
						"format": "double"
					},
					"eventId": {
						"type": "number",
						"format": "double"
					},
					"eventName": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"startDate": {
						"type": "string"
					},
					"endDate": {
						"type": "string"
					},
					"cost": {
						"items": {
							"$ref": "#/components/schemas/Cost"
						},
						"type": "array"
					}
				},
				"required": [
					"tripId",
					"eventId",
					"eventName",
					"location",
					"startDate",
					"endDate",
					"cost"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_TripEventResponse_": {
				"description": "통일된 API 응답 형식",
				"properties": {
					"isSuccess": {
						"type": "boolean"
					},
					"code": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"allOf": [
							{
								"$ref": "#/components/schemas/TripEventResponse"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"isSuccess",
					"code",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateTripEventRequest": {
				"properties": {
					"tripId": {
						"type": "number",
						"format": "double"
					},
					"eventName": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"startDate": {
						"type": "string"
					},
					"endDate": {
						"type": "string"
					},
					"cost": {
						"items": {
							"$ref": "#/components/schemas/Cost"
						},
						"type": "array"
					}
				},
				"required": [
					"tripId",
					"eventName",
					"location",
					"startDate",
					"endDate",
					"cost"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TripEventRequest": {
				"properties": {
					"tripId": {
						"type": "number",
						"format": "double"
					},
					"eventId": {
						"type": "number",
						"format": "double"
					},
					"eventName": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"startDate": {
						"type": "string"
					},
					"endDate": {
						"type": "string"
					},
					"cost": {
						"items": {
							"$ref": "#/components/schemas/Cost"
						},
						"type": "array"
					}
				},
				"required": [
					"tripId",
					"eventName",
					"location",
					"startDate",
					"endDate",
					"cost"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_TripEventResponse-Array_": {
				"description": "통일된 API 응답 형식",
				"properties": {
					"isSuccess": {
						"type": "boolean"
					},
					"code": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"items": {
							"$ref": "#/components/schemas/TripEventResponse"
						},
						"type": "array",
						"nullable": true
					}
				},
				"required": [
					"isSuccess",
					"code",
					"message",
					"result"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "access_token",
				"in": "cookie"
			}
		}
	},
	"info": {
		"title": "server",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/trips": {
			"post": {
				"operationId": "CreateTrip",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TripScheduleResponse_"
								}
							}
						}
					}
				},
				"description": "여행 일정 생성",
				"tags": [
					"TripSchedule"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTripRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteTrips",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "여러 여행 일정 삭제",
				"tags": [
					"TripSchedule"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"ids": {
										"items": {
											"type": "number",
											"format": "double"
										},
										"type": "array"
									}
								},
								"required": [
									"ids"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
		"/trips/{tripId}": {
			"get": {
				"operationId": "GetTripById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TripScheduleResponse_"
								}
							}
						}
					}
				},
				"description": "단일 여행 데이터 조회",
				"tags": [
					"TripSchedule"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tripId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateTripSchedule",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TripScheduleResponse_"
								}
							}
						}
					}
				},
				"description": "여행 일정 업데이트",
				"tags": [
					"TripSchedule"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tripId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateTripRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteTrip",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "여행 일정 삭제",
				"tags": [
					"TripSchedule"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "tripId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/event": {
			"post": {
				"operationId": "CreateTripEvent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TripEventResponse_"
								}
							}
						}
					}
				},
				"description": "이벤트 생성",
				"tags": [
					"TripEvent"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTripEventRequest"
							}
						}
					}
				}
			}
		},
		"/event/{event_id}": {
			"patch": {
				"operationId": "UpdateTripEvent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TripEventResponse_"
								}
							}
						}
					}
				},
				"description": "이벤트 업데이트",
				"tags": [
					"TripEvent"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "event_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TripEventRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetTripEventById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TripEventResponse_"
								}
							}
						}
					}
				},
				"description": "event_id로 event 단일 조회",
				"tags": [
					"TripEvent"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "event_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteTripEventById",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "event_id로 event 단일 삭제",
				"tags": [
					"TripEvent"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "event_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/event/all/{trip_id}": {
			"get": {
				"operationId": "GetTripEventsByTripId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TripEventResponse-Array_"
								}
							}
						}
					}
				},
				"description": "trip_id로 event 모두 조회",
				"tags": [
					"TripEvent"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "trip_id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}